deploy-send-email-main/requirements.txt
fastapi==0.115.1
uvicorn==0.23.2
requests==2.32.3
python-dotenv==1.0.1
pydantic[email]==2.10.6
